Revert "mgr/dashboard: set up dashboard as a app shell"
Fixes https://tracker.ceph.com/issues/74006

This reverts commit a0dd52fe100932922ceab9277490bfa2f8631431.

 Conflicts:
	src/pybind/mgr/dashboard/frontend/module-federation.config.ts
	src/pybind/mgr/dashboard/frontend/package-lock.json
	src/pybind/mgr/dashboard/frontend/package.json
	src/pybind/mgr/dashboard/frontend/project.json
