+++ src/pybind/mgr/dashboard/frontend/.gitignore
@@ -48,9 +48,3 @@ yarn.lock
 # Ceph
 !core
 !*.core
-
-
-
-.nx/cache
-.nx/workspace-data
-
