+++ src/pybind/mgr/dashboard/frontend/package.json
@@ -6,20 +6,25 @@
     "locale": "en-US"
   },
   "scripts": {
-    "start": "npm run env_build && nx serve --host 0.0.0.0 --ssl",
-    "build": "npm run env_build && nx build --skip-nx-cache --configuration=$npm_package_config_locale",
-    "build:localize": "node cd --env --pre && nx build --localize",
+    "ng": "ng",
+    "start": "npm run env_build && ng serve --host 0.0.0.0 --ssl",
+    "build": "npm run env_build && ng build --configuration=$npm_package_config_locale",
+    "build:localize": "node cd --env --pre && ng build --localize",
     "postbuild:localize": "node cd --res",
     "env_build": "node cd --env",
     "i18n": "npm run i18n:extract",
-    "i18n:extract": "nx extract-i18n --output-path src/locale --progress=false",
+    "i18n:extract": "ng extract-i18n --output-path src/locale --progress=false",
+    "i18n:push": "npx i18ntool push -c i18n.config.json",
+    "i18n:pull": "npx i18ntool pull -c i18n.config.json",
+    "i18n:merge": "npx i18ntool merge -c i18n.config.json",
+    "i18n:token": "npx i18ntool config token",
     "test": "jest --watch",
     "test:ci": "jest --clearCache && JEST_SILENT_REPORTER_DOTS=true jest --detectOpenHandles --coverage --reporters jest-silent-reporter",
     "pree2e": "rm -f cypress/reports/results-*.xml || true",
     "e2e": "start-test 4200 'cypress open'",
     "pree2e:ci": "npm run pree2e",
     "e2e:ci": "start-test 4200 'cypress run -b chrome --headless'",
-    "lint:eslint": "nx lint",
+    "lint:eslint": "ng lint",
     "lint:gherkin": "gherkin-lint -c .gherkin-lintrc cypress/e2e",
     "lint:prettier": "prettier --list-different \"{src,cypress}/**/*.{ts,scss}\"",
     "lint:html": "htmllint src/app/**/*.html && html-linter --config html-linter.config.json",
@@ -54,7 +59,6 @@
     "@ng-bootstrap/ng-bootstrap": "17.0.1",
     "@ngx-formly/bootstrap": "6.1.1",
     "@ngx-formly/core": "6.1.1",
-    "@nx/angular": "20.8.1",
     "@popperjs/core": "2.11.8",
     "@types/file-saver": "2.0.7",
     "async-mutex": "0.5.0",
@@ -98,12 +102,6 @@
     "@compodoc/compodoc": "1.1.26",
     "@cypress/browserify-preprocessor": "3.0.2",
     "@juggle/resize-observer": "3.4.0",
-    "@module-federation/enhanced": "^0.9.0",
-    "@nx/module-federation": "20.8.1",
-    "@nx/webpack": "20.8.1",
-    "@nx/workspace": "20.8.1",
-    "@rspack/core": "1.4.11",
-    "@schematics/angular": "18.2.11",
     "@types/brace-expansion": "1.1.0",
     "@types/cypress": "0.1.6",
     "@types/cypress-axe": "0.8.0",
@@ -137,7 +135,6 @@
     "mocha-junit-reporter": "2.1.0",
     "ng-mocks": "14.13.1",
     "npm-run-all": "4.1.5",
-    "nx": "20.8.1",
     "postcss-scss": "4.0.9",
     "prettier": "2.1.2",
     "pretty-quick": "3.0.2",
