+++ src/pybind/mgr/dashboard/frontend/src/app/app.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { provideRouter } from '@angular/router';
+import { RouterTestingModule } from '@angular/router/testing';
 
 import { configureTestBed } from '~/testing/unit-test-helper';
 import { AppComponent } from './app.component';
@@ -10,7 +10,7 @@ describe('AppComponent', () => {
 
   configureTestBed({
     declarations: [AppComponent],
-    providers: [provideRouter([])]
+    imports: [RouterTestingModule]
   });
 
   beforeEach(() => {
