--- src/pybind/mgr/dashboard/frontend/src/assets/module-federation.manifest.json
@@ -1,1 +1,0 @@
-{}
