+++ src/pybind/mgr/dashboard/frontend/tsconfig.json
@@ -22,9 +22,13 @@
     "baseUrl": "./",
     "resolveJsonModule": true,
     "paths": {
-      "stream": ["node_modules/stream-browserify"],
-      "timers": ["node_modules/timers-browserify"],
-      "~/*": ["src/*"]
+      "~/*": ["src/*"],
+      "timers": [
+          "node_modules/timers-browserify"
+        ],
+        "stream": [
+          "node_modules/stream-browserify"
+        ]
     },
     "typeRoots": ["node_modules/@types"],
     "lib": ["es2017", "dom"],
