--- src/pybind/mgr/dashboard/frontend/webpack.config.ts
@@ -1,9 +1,0 @@
-import { withModuleFederation } from '@nx/module-federation/angular';
-import config from './module-federation.config';
-
-/**
- * DTS Plugin is disabled in Nx Workspaces as Nx already provides Typing support for Module Federation
- * The DTS Plugin can be enabled by setting dts: true
- * Learn more about the DTS Plugin here: https://module-federation.io/configure/dts.html
- */
-export default withModuleFederation(config, { dts: false });
